Move `ProtectedState` to its own file
