+++ app/src/main/java/org/simple/clinic/security/pin/BruteForceProtection.kt
@@ -12,38 +12,12 @@ import org.simple.clinic.util.timer
 import org.threeten.bp.Duration
 import org.threeten.bp.Instant
 import javax.inject.Inject
-import kotlin.math.max
 
 class BruteForceProtection @Inject constructor(
     private val utcClock: UtcClock,
     private val config: BruteForceProtectionConfig,
     private val statePreference: Preference<BruteForceProtectionState>
 ) {
-
-  sealed class ProtectedState {
-
-    companion object {
-      fun from(
-          state: BruteForceProtectionState,
-          maxAllowedFailedAttempts: Int,
-          blockAttemptsFor: Duration
-      ): ProtectedState {
-        val blockedAt = state.limitReachedAt
-        val attemptsMade = state.failedAuthCount
-
-        return when (blockedAt) {
-          is None -> {
-            val attemptsRemaining = max(0, maxAllowedFailedAttempts - attemptsMade)
-            Allowed(attemptsMade = attemptsMade, attemptsRemaining = attemptsRemaining)
-          }
-          is Just -> Blocked(attemptsMade = attemptsMade, blockedTill = blockedAt.value + blockAttemptsFor)
-        }
-      }
-    }
-
-    data class Allowed(val attemptsMade: Int, val attemptsRemaining: Int) : ProtectedState()
-    data class Blocked(val attemptsMade: Int, val blockedTill: Instant) : ProtectedState()
-  }
 
   fun incrementFailedAttempt(): Completable {
     return Completable.fromAction {
